% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{accuracy}
\alias{accuracy}
\title{Calculate Accuracy of binary classification}
\usage{
accuracy(label, pred)
}
\arguments{
\item{label}{True labels}

\item{pred}{predicted labels}
}
\description{
Calculate Accuracy of binary classification
}
