% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC.R
\name{AUC}
\alias{AUC}
\title{Calculate AUC of binary classification}
\usage{
AUC(label, pred)
}
\arguments{
\item{label}{true labels}

\item{pred}{predicted lables}
}
\value{
AUC
}
\description{
Calculate AUC of binary classification
}
