% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BARTr_train.R
\name{BARTr_train}
\alias{BARTr_train}
\title{BART model train workhourse}
\usage{
BARTr_train(X, Rj, treej, p_modify, Tmin, rule, sigma2, tau, base, power)
}
\arguments{
\item{rule}{grp: Gaussian random projection; sgrp: sparse Gaussian random projection; bart: originla bart; hyperplane: connect two points}
}
\description{
BART model train workhourse
}
