% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{change_tree}
\alias{change_tree}
\title{Change proposal}
\usage{
change_tree(btree_obj, X, Tmin, rule)
}
\value{
changed tree, old leaf mode data index, new leaf node data index
}
\description{
Change proposal
}
