% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_terminal_idx.R
\name{find_terminal_idx}
\alias{find_terminal_idx}
\title{Find which leaf node a test sample belongs to}
\usage{
find_terminal_idx(X_test, btree_obj)
}
\description{
Find which leaf node a test sample belongs to
}
