% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gen_R.R
\name{Gen_R}
\alias{Gen_R}
\title{Generate an array of random rotation matrices}
\usage{
Gen_R(p, nr)
}
\arguments{
\item{p}{p*p matrix}

\item{nr}{number of matrices}
}
\description{
Generate an array of random rotation matrices
}
