% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gramschmidt.R
\name{gramschmidt}
\alias{gramschmidt}
\title{QR Decomposition via gram schmidt method}
\usage{
gramschmidt(x)
}
\description{
This is an R function from \url{https://rpubs.com/aaronsc32/qr-decomposition-gram-schmidt}
}
