% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow.R
\name{grow_tree}
\alias{grow_tree}
\title{Grow tree proposal}
\usage{
grow_tree(btree_obj, X, Tmin, rule)
}
\value{
growed tree, old tree depth, old tree leaf data index, new tree leaf data index.
}
\description{
Grow tree proposal
}
