% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik_ratio_change.R
\name{lik_ratio_change}
\alias{lik_ratio_change}
\title{Evaluate liklihood ratio of new tree and old tree for CHANGE}
\usage{
lik_ratio_change(old_t_data, new_t_data, Rj, Tmin, sigma2, tau)
}
\description{
Evaluate liklihood ratio of new tree and old tree for CHANGE
}
