% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik_ratio_grow.R
\name{lik_ratio_grow}
\alias{lik_ratio_grow}
\title{Evaluate liklihood ratio of new tree and old tree for GROW}
\usage{
lik_ratio_grow(old_t_data, new_t_data, Rj, Tmin, sigma2, tau)
}
\arguments{
\item{old_t_data}{a list of old tree leaf data index}

\item{new_t_data}{a list of new tree leaf data index}
}
\description{
Evaluate liklihood ratio of new tree and old tree for GROW
}
