% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik_ratio_prune.R
\name{lik_ratio_prune}
\alias{lik_ratio_prune}
\title{Evaluate liklihood ratio of new tree and old tree for PRUNE}
\usage{
lik_ratio_prune(old_t_data, new_t_data, Rj, Tmin, sigma2, tau)
}
\description{
Evaluate liklihood ratio of new tree and old tree for PRUNE
}
