% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_split.R
\name{new_split}
\alias{new_split}
\title{Genrate new splitted data index}
\usage{
new_split(X_sub, sub_data, dir_new, rule_new)
}
\value{
left and right data index
}
\description{
Genrate new splitted data index
}
