% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pBADS.R
\name{pBADS}
\alias{pBADS}
\title{Classification BART}
\usage{
pBADS(X, y, x.test, cutoff = 0.5, k = 2, binaryOffset = NULL,
  ntree = 50, ndpost = 700, nskip = 300, Tmin = 2,
  printevery = 100, save_trees = F, rule = "bart", pre_train = T)
}
\arguments{
\item{rule}{grp: Gaussian random projection; sgrp: sparse Gaussian random projection; bart: originla bart; hyperplane: connect two points}
}
\description{
Classification BART
}
