% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pBARTr_CV.R
\name{pBARTr_CV}
\alias{pBARTr_CV}
\title{pBART cross-validation}
\usage{
pBARTr_CV(x, y, ks = c(0.25, 0.5, 1, 2, 3), ntrees = c(50, 100),
  powers = c(2, 3, 5, 10), bases = c(0.5, 0.8, 0.95), rs = c(2, 3, 5,
  10, 20), p_split, nfold = 5)
}
\arguments{
\item{ks, powers, bases, ntrees, rs, p_split}{ks: a vector of k; so on}

\item{nfold}{number of folds}
}
\description{
pBART cross-validation
}
