% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RotBART.R
\name{RotBART}
\alias{RotBART}
\title{Rotation BART for regression}
\usage{
RotBART(X, y, x.test, sigdf = 3, sigquant = 0.9, k = 2,
  lambda = NA, sigest = NA, sigmaf = NA, power = 2, base = 0.95,
  ntree = 50, ndpost = 700, nskip = 300, Tmin = 2,
  printevery = 100, p_modify = c(2.5, 2.5, 4)/9, save_trees = F,
  rule = "bart", p_split = "CGM", rotate = "rr", srpk = 2 *
  ncol(X))
}
\arguments{
\item{rotate}{'rr'='random rotation'','rraug'='random rotation+augmentation','srp'='sparse random projection'}

\item{srpk}{the number of cols of sparse projection matrix}

\item{Others}{See ?BARTr}
}
\value{
See ?BARTr
}
\description{
Rotation BART for regression
}
