% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RotpBADS.R
\name{RotpBADS}
\alias{RotpBADS}
\title{Rotation BADS for classification}
\usage{
RotpBADS(X, y, x.test, cutoff = 0.5, k = 2, binaryOffset = NULL,
  ntree = 50, ndpost = 700, nskip = 300, Tmin = 2,
  printevery = 100, save_trees = F, rotate = "rr", srpk = 2 *
  ncol(X))
}
\arguments{
\item{rotate}{'rr'='random rotation'','rraug'='random rotation+augmentation','srp'='sparse random projection'}

\item{srpk}{the number of cols of sparse projection matrix}
}
\description{
Rotation BADS for classification
}
