% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RotpBART.R
\name{RotpBART}
\alias{RotpBART}
\title{Rotation BART for classification}
\usage{
RotpBART(X, y, x.test, cutoff = 0.5, k = 2, binaryOffset = NULL,
  power = 2, base = 0.95, w = rep(1, length(y)), ntree = 50,
  ndpost = 700, nskip = 300, Tmin = 2, printevery = 100,
  p_modify = c(2.5, 2.5, 4)/9, save_trees = F, rule = "bart",
  p_split = "CGM", rotate = "rr", srpk = 2 * ncol(X))
}
\arguments{
\item{rotate}{'rr'='random rotation'','rraug'='random rotation+augmentation','srp'='sparse random projection'}

\item{srpk}{the number of cols of sparse projection matrix}

\item{Others}{See ?BARTr}
}
\value{
See ?BARTr
}
\description{
Rotation BART for classification
}
