% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yhat_draw_bads.R
\name{yhat.draw.bads}
\alias{yhat.draw.bads}
\title{Draw of y when accept tree movement for bads version}
\usage{
yhat.draw.bads(btree_obj, x.test, Rj, tau, sigma2, rotate)
}
\value{
Draws of yhat and ypred
}
\description{
Draw of y when accept tree movement for bads version
}
