% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yhat_draw.R
\name{yhat.draw}
\alias{yhat.draw}
\title{Draw of y when accept tree movement}
\usage{
yhat.draw(btree_obj, x.test, Rj, tau, sigma2, cl)
}
\value{
Draws of yhat and ypred
}
\description{
Draw of y when accept tree movement
}
