% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yhat_draw_rotation.R
\name{yhat.draw.rotation}
\alias{yhat.draw.rotation}
\title{Draw of y when accept tree movement for rotation version}
\usage{
yhat.draw.rotation(btree_obj, x.test, Rj, tau, sigma2, rotate)
}
\value{
Draws of yhat and ypred
}
\description{
Draw of y when accept tree movement for rotation version
}
