% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yhat_draw_train.R
\name{yhat.draw.train}
\alias{yhat.draw.train}
\title{Draw of y training sample only}
\usage{
yhat.draw.train(btree_obj, Rj, tau, sigma2)
}
\value{
Draws of yhat
}
\description{
Draw of y training sample only
}
