% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yhat_draw2_bads.R
\name{yhat.draw2.bads}
\alias{yhat.draw2.bads}
\title{Draw of y when reject tree movement for bads version}
\usage{
yhat.draw2.bads(btree_obj, x.test, Rj, tau, sigma2, rotate)
}
\value{
Draws of yhat and ypred
}
\description{
Draw of y when reject tree movement for bads version
}
