% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yhat_draw2.R
\name{yhat.draw2}
\alias{yhat.draw2}
\title{Draw of y when reject tree movement}
\usage{
yhat.draw2(btree_obj, x.test, Rj, tau, sigma2)
}
\value{
Draws of yhat and ypred
}
\description{
Draw of y when reject tree movement
}
