% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yhat_draw2_rotation.R
\name{yhat.draw2.rotation}
\alias{yhat.draw2.rotation}
\title{Draw of y when reject tree movement for rotation version}
\usage{
yhat.draw2.rotation(btree_obj, x.test, Rj, tau, sigma2, rotate)
}
\value{
Draws of yhat and ypred
}
\description{
Draw of y when reject tree movement for rotation version
}
