% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gen_SRP.R
\name{Gen_SRP}
\alias{Gen_SRP}
\title{Generate an array of sparse random orojection matrices}
\usage{
Gen_SRP(p, k, nr, lb = 1/3)
}
\arguments{
\item{p, k}{matrix dimension p*k}

\item{nr}{number of matrices}

\item{lb}{prop of 0's in matrix}
}
\description{
Generate an array of sparse random orojection matrices
}
