% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbart_reps.R
\name{pbart_reps}
\alias{pbart_reps}
\title{Run reps of pbart on real data set}
\usage{
pbart_reps(x, y, ntrain, nmaxtest = 1000, nreps = 50, seed = 12345,
  ntrees = c(50, 200), ks = c(0.5, 2, 5), powers = c(2, 3, 5, 10,
  ntrain), k_folds = 5)
}
\description{
Run reps of pbart on real data set
}
