% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pBARTr.R
\name{pBARTr}
\alias{pBARTr}
\title{Classification BART}
\usage{
pBARTr(X, y, x.test, cutoff = 0.5, k = 2, binaryOffset = NULL,
  power = 2, base = 0.95, w = rep(1, length(y)), ntree = 50,
  ndpost = 700, nskip = 300, Tmin = 5, printevery = 100,
  p_modify = c(2.5, 2.5, 4)/9, save_trees = F, rule = "bart")
}
\arguments{
\item{rule}{grp: Gaussian random projection; sgrp: sparse Gaussian random projection; bart: originla bart; hyperplane: connect two points}
}
\description{
Classification BART
}
