% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RotpBADS.R
\name{RotpBADS}
\alias{RotpBADS}
\title{Rotation BADS for classification}
\usage{
RotpBADS(X, y, x.test, cutoff = 0.5, k = 2, binaryOffset = NULL,
  ntree = 50, ndpost = 700, nskip = 300, Tmin = 2,
  printevery = 100, save_trees = F, rule = "bart", rotate = "rr")
}
\arguments{
\item{rotate}{'rr'='random rotation'','rraug'='random rotation+augmentation'}

\item{Others}{see ?pBADS}
}
\value{
See ?pBADS
}
\description{
Rotation BADS for classification
}
